core: move public_key_to_wif to helpers
