+++ core/src/apps/eos/helpers.py
@@ -1,3 +1,4 @@
+from trezor import wire
 from trezor.crypto import base58
 from trezor.messages.EosAsset import EosAsset
 
@@ -61,3 +62,14 @@ def validate_full_path(path: list) -> bool:
     if path[4] != 0:
         return False
     return True
+
+
+def public_key_to_wif(pub_key: bytes) -> str:
+    if pub_key[0] == 0x04 and len(pub_key) == 65:
+        head = b"\x03" if pub_key[64] & 0x01 else b"\x02"
+        compressed_pub_key = head + pub_key[1:33]
+    elif pub_key[0] in [0x02, 0x03] and len(pub_key) == 33:
+        compressed_pub_key = pub_key
+    else:
+        raise wire.DataError("invalid public key")
+    return base58_encode("EOS", "", compressed_pub_key)
