+++ core/tests/test_apps.eos.get_public_key.py
@@ -4,8 +4,8 @@ from trezor.crypto import bip32, bip39
 from apps.common.paths import HARDENED
 
 if not utils.BITCOIN_ONLY:
-    from apps.eos.get_public_key import _get_public_key, public_key_to_wif
-    from apps.eos.helpers import validate_full_path
+    from apps.eos.get_public_key import _get_public_key
+    from apps.eos.helpers import validate_full_path, public_key_to_wif
 
 
 @unittest.skipUnless(not utils.BITCOIN_ONLY, "altcoin")
